+++ src/main/java/com/smartsheet/api/ReportResources.java
@@ -21,12 +21,14 @@ import com.smartsheet.api.models.PaginationParameters;
 import com.smartsheet.api.models.Report;
 import com.smartsheet.api.models.ReportDefinition;
 import com.smartsheet.api.models.ReportPublish;
+import com.smartsheet.api.models.ReportScopeInclusion;
 import com.smartsheet.api.models.SheetEmail;
 import com.smartsheet.api.models.enums.ReportInclusion;
 
 import java.io.OutputStream;
 import java.util.Date;
 import java.util.EnumSet;
+import java.util.List;
 
 /**
  * <p>This interface provides methods to access Report resources.</p>
@@ -242,4 +244,51 @@ public interface ReportResources {
      * @return the created ShareResources object
      */
     ShareResources shareResources();
+
+    /**
+     * <p>Deletes a report.</p>
+     *
+     * <p>Mirrors the following Smartsheet REST API method: DELETE /reports/{reportId}</p>
+     *
+     * @param id the id of the report
+     * @throws IllegalArgumentException    if any argument is null or empty string
+     * @throws InvalidRequestException     if there is any problem with the REST API request
+     * @throws AuthorizationException      if there is any problem with  the REST API authorization (access token)
+     * @throws ResourceNotFoundException   if the resource cannot be found
+     * @throws ServiceUnavailableException if the REST API service is not available (possibly due to rate limiting)
+     * @throws SmartsheetException         if there is any other error during the operation
+     */
+    void deleteReport(long id) throws SmartsheetException;
+    
+    /**
+     * <p>Adds one or more specified sheet or workspace to the report scope.</p>
+     *
+     * @param id            the ID of the report
+     * @param scopes   A list of one or more objects denoting the sheets or workspaces associated with
+     *                 the report to be added to the report scope.
+     *
+     * @throws IllegalArgumentException    if any argument is null or empty string
+     * @throws InvalidRequestException     if there is any problem with the REST API request
+     * @throws AuthorizationException      if there is any problem with  the REST API authorization (access token)
+     * @throws ResourceNotFoundException   if the resource cannot be found
+     * @throws ServiceUnavailableException if the REST API service is not available (possibly due to rate limiting)
+     * @throws SmartsheetException         if there is any other error during the operation
+     */
+    void addReportScope(long id, List<ReportScopeInclusion> scopes) throws SmartsheetException;
+
+    /**
+     * <p>Removes one or more specified sheet or workspace from the report scope.</p>
+     *
+     * @param id                the ID of the report
+     * @param scopes    A list of one or more objects denoting the sheets or workspaces associated with
+     *                  the report to be removed from the report scope.
+     *
+     * @throws IllegalArgumentException    if any argument is null or empty string
+     * @throws InvalidRequestException     if there is any problem with the REST API request
+     * @throws AuthorizationException      if there is any problem with  the REST API authorization (access token)
+     * @throws ResourceNotFoundException   if the resource cannot be found
+     * @throws ServiceUnavailableException if the REST API service is not available (possibly due to rate limiting)
+     * @throws SmartsheetException         if there is any other error during the operation
+     */
+    void removeReportScope(long id, List<ReportScopeInclusion> scopes) throws SmartsheetException;
 }