+++ src/main/java/com/smartsheet/api/models/enums/ReportAssetType.java
@@ -1,0 +1,32 @@
+/*
+ * Copyright (C) 2025 Smartsheet
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.smartsheet.api.models.enums;
+
+/**
+ * Represents the type of asset that is included in the report.
+ */
+public enum ReportAssetType {
+    /**
+     * A sheet that is included in the report.
+     */
+    SHEET,
+
+    /**
+     * A workspace that is included in the report.
+     */
+    WORKSPACE
+}